Day 12: part 1 (there is no part 2)
If fails for the example input, but works for the full input, because the full input is simpler.

For the full input, we never have to actually try and make the shapes interlock.  Either the areas of the shapes total more than the area of the region (trivially won't fit), or the region can be broken down into enough 3x3 sub-regions where each shape can occupy its own sub-region.